Вынос инициализации контейнера вне точки входа
